swapspace/requirements.txt
aioredis==1.2.0
amqp==2.4.2
asgiref==3.1.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
channels==2.1.7
channels-redis==2.3.3
chardet==3.0.4
constantly==15.1.0
daphne==2.3.0
Django==2.1.4
django-extensions==2.1.6
django-paypal==1.0.0
djangorestframework==3.9.2
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
kombu==4.5.0
msgpack==0.6.1
Pillow==5.4.1
PyHamcrest==1.9.0
pytz==2019.1
redis==3.2.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.2.0
txaio==18.8.1
urllib3==1.24.2
vine==1.3.0
whitenoise==4.1.2
zope.interface==4.6.0
